# api/openapi.yaml
openapi: 3.1.0
info:
  title: ClimaControl API
  version: 1.0.0
  description: |
    API REST do ClimaControl — gateway sobre o Firebase Realtime Database.
    Projeto Integrador UNIFECAF.

    **Contrato como fonte da verdade (design-first):** este `openapi.yaml` é
    carregado pela aplicação em tempo de execução e servido no Swagger UI
    (`/api/docs`). Validação de entrada por JSON Schema/Zod; erros no padrão
    **RFC 7807 (Problem Details)**.
  contact:
    name: Equipe ClimaControl
  license:
    name: MIT

servers:
  - url: https://climacontrol-api.onrender.com
    description: Produção (Render)
  - url: http://localhost:3001
    description: Desenvolvimento local

tags:
  - name: sensores
    description: Leitura atual dos sensores (recurso singleton)
  - name: controle
    description: Estado de controle do A/C (recurso singleton)
  - name: historico
    description: Série temporal (coleção, paginada)
  - name: eventos
    description: Log de eventos (coleção, paginada)

security: []

paths:
  /api/v1/sensores:
    get:
      tags: [sensores]
      summary: Leitura atual dos sensores
      description: Retorna a leitura corrente (ocupação, temperaturas interna/externa e umidades).
      operationId: getSensores
      responses:
        "200":
          description: Leitura atual
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SensoresEnvelope" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
    put:
      tags: [sensores]
      summary: ESP32 envia a leitura completa (substitui o estado)
      description: PUT (idempotente) — o dispositivo envia o objeto inteiro a cada ciclo.
      operationId: putSensores
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SensoresInput" }
      responses:
        "200":
          description: Atualizado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SensoresEnvelope" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }

  /api/v1/controle:
    get:
      tags: [controle]
      summary: Estado atual de controle do A/C
      description: Retorna o estado de controle (ligado/desligado, modo manual, temperatura alvo).
      operationId: getControle
      responses:
        "200":
          description: Estado de controle
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ControleEnvelope" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
    patch:
      tags: [controle]
      summary: Atualização parcial (ligar A/C, ajustar temp_alvo, modo)
      description: |
        PATCH (e não PUT) porque o cliente envia só os campos que mudaram.
        Regra de negócio: `temp_alvo` ∈ [16, 28].
      operationId: patchControle
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ControlePatch" }
            examples:
              ligar:
                summary: Ligar e ajustar alvo
                value: { ac_ligado: true, temp_alvo: 22 }
      responses:
        "200":
          description: Estado atualizado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ControleEnvelope" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }

  /api/v1/historico:
    get:
      tags: [historico]
      summary: Série temporal filtrada e paginada
      description: Lista as amostras do histórico dentro da janela de tempo (period), paginadas e ordenadas.
      operationId: listHistorico
      parameters:
        - name: period
          in: query
          schema: { type: string, enum: ["1h", "4h", "8h", "12h", "24h"], default: "12h" }
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - name: sort
          in: query
          schema: { type: string, enum: [asc, desc], default: asc }
      responses:
        "200":
          description: Página do histórico
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HistoricoPage" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
    post:
      tags: [historico]
      summary: ESP32 anexa uma amostra
      description: Anexa uma amostra de telemetria ao histórico (uso do dispositivo ESP32).
      operationId: createHistorico
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/HistoricoSample" }
      responses:
        "201":
          description: Amostra criada
          headers:
            Location:
              schema: { type: string }
              description: Caminho do recurso criado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HistoricoSampleEnvelope" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }

  /api/v1/eventos:
    get:
      tags: [eventos]
      summary: Lista paginada de eventos
      description: Lista eventos de controle/auditoria, paginados, com filtro por tipo e ordenação.
      operationId: listEventos
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - name: sort
          in: query
          schema: { type: string, enum: [asc, desc], default: desc }
        - name: type
          in: query
          description: Filtro opcional por tipo de evento
          schema: { type: string, maxLength: 64 }
      responses:
        "200":
          description: Página de eventos
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EventosPage" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
    post:
      tags: [eventos]
      summary: Registra um evento
      description: Registra um novo evento de controle/auditoria na coleção.
      operationId: createEvento
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EventoInput" }
      responses:
        "201":
          description: Evento criado
          headers:
            Location:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EventoEnvelope" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }

  /api/v1/eventos/{id}:
    delete:
      tags: [eventos]
      summary: Remove um evento (idempotente)
      description: Remove um evento pelo id. Idempotente — repetir após remover devolve 404.
      operationId: deleteEvento
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204": { description: Removido (sem corpo) }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        ID token (JWT) do Firebase Authentication no header
        `Authorization: Bearer <token>`. Validado no servidor via
        Admin SDK (`verifyIdToken`).

  parameters:
    Page:
      name: page
      in: query
      description: Página (1-based)
      schema: { type: integer, minimum: 1, default: 1 }
    Limit:
      name: limit
      in: query
      description: Itens por página (teto de 100)
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }

  responses:
    NotFound:
      description: Recurso não encontrado
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Unauthorized:
      description: Autenticação necessária ou token inválido
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    ValidationError:
      description: Entrada inválida
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
          example:
            type: https://climacontrol-api.onrender.com/problems/validation-error
            title: Dados de entrada inválidos.
            status: 400
            detail: Um ou mais campos da requisição são inválidos.
            instance: /api/v1/controle
            errors:
              - campo: temp_alvo
                mensagem: temp_alvo deve ser no máximo 28°C
    ServiceUnavailable:
      description: Dependência (Firebase) indisponível
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    TooManyRequests:
      description: Limite de requisições excedido (rate limiting)
      headers:
        Retry-After:
          schema: { type: integer }
          description: Segundos a aguardar antes de tentar novamente.
        RateLimit-Remaining:
          schema: { type: integer }
          description: Requisições restantes na janela atual.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }

  schemas:
    Meta:
      type: object
      properties:
        version: { type: string, example: v1 }
        timestamp: { type: integer, example: 1730000000000 }

    Pagination:
      type: object
      properties:
        page: { type: integer, example: 1 }
        page_size: { type: integer, example: 50 }
        total: { type: integer, example: 120 }
        total_pages: { type: integer, example: 3 }
        has_more: { type: boolean, example: true }

    Problem:
      type: object
      description: Erro no formato RFC 7807 (Problem Details).
      properties:
        type: { type: string, format: uri, example: "https://climacontrol-api.onrender.com/problems/not-found" }
        title: { type: string, example: Recurso não encontrado. }
        status: { type: integer, example: 404 }
        detail: { type: string }
        instance: { type: string, example: /api/v1/sensores }
        errors:
          type: array
          items:
            type: object
            properties:
              campo: { type: string }
              mensagem: { type: string }
      required: [type, title, status]

    Sensores:
      type: object
      properties:
        ocupacao: { type: integer, example: 12 }
        temp_interna: { type: number, example: 23.5 }
        temp_externa: { type: number, example: 31.2 }
        temp_alvo: { type: integer, example: 22 }
        umid_interna: { type: number, example: 55 }
        umid_externa: { type: number, example: 70 }
        ac_ligado: { type: boolean, example: true }
        modo_manual: { type: boolean, example: false }
    SensoresInput:
      type: object
      required: [ocupacao, temp_interna, temp_externa, temp_alvo, umid_interna, umid_externa, ac_ligado]
      properties:
        ocupacao: { type: integer, minimum: 0, maximum: 500 }
        temp_interna: { type: number }
        temp_externa: { type: number }
        temp_alvo: { type: integer, minimum: 0, maximum: 30 }
        umid_interna: { type: number }
        umid_externa: { type: number }
        ac_ligado: { type: boolean }
    SensoresEnvelope:
      type: object
      properties:
        data: { $ref: "#/components/schemas/Sensores" }
        meta: { $ref: "#/components/schemas/Meta" }

    Controle:
      type: object
      properties:
        ac_ligado: { type: boolean }
        modo_manual: { type: boolean }
        temp_alvo: { type: integer }
        ia_ativa: { type: boolean }
        atualizado_em: { type: integer }
    ControlePatch:
      type: object
      minProperties: 1
      description: Ao menos um campo deve ser enviado.
      properties:
        ac_ligado: { type: boolean }
        modo_manual: { type: boolean }
        temp_alvo: { type: integer, minimum: 16, maximum: 28, example: 22 }
        ia_ativa: { type: boolean }
    ControleEnvelope:
      type: object
      properties:
        data: { $ref: "#/components/schemas/Controle" }
        meta: { $ref: "#/components/schemas/Meta" }

    HistoricoSample:
      type: object
      required: [o, ti, te, ta, ui, ue]
      properties:
        t: { type: integer, description: "timestamp ms (servidor preenche se ausente)" }
        o: { type: integer, description: ocupação }
        ti: { type: number, description: temp interna }
        te: { type: number, description: temp externa }
        ta: { type: integer, description: temp alvo }
        ui: { type: number, description: umidade interna }
        ue: { type: number, description: umidade externa }
    HistoricoSampleEnvelope:
      type: object
      properties:
        data: { $ref: "#/components/schemas/HistoricoSample" }
        meta: { $ref: "#/components/schemas/Meta" }
    HistoricoPage:
      type: object
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/HistoricoSample" }
        meta:
          allOf:
            - $ref: "#/components/schemas/Meta"
            - type: object
              properties:
                period: { type: string, example: 4h }
                pagination: { $ref: "#/components/schemas/Pagination" }

    Evento:
      type: object
      properties:
        id: { type: string, example: "-OtuzkqS45WjcBwZfqSd" }
        type: { type: string, example: ac_ligado_manual }
        payload: {}
        timestamp: { type: integer }
    EventoInput:
      type: object
      required: [type]
      properties:
        type: { type: string, maxLength: 64, example: ac_ligado_manual }
        payload: {}
        timestamp: { type: integer }
    EventoEnvelope:
      type: object
      properties:
        data: { $ref: "#/components/schemas/Evento" }
        meta: { $ref: "#/components/schemas/Meta" }
    EventosPage:
      type: object
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/Evento" }
        meta:
          allOf:
            - $ref: "#/components/schemas/Meta"
            - type: object
              properties:
                pagination: { $ref: "#/components/schemas/Pagination" }